ts_less/less_ts: avoid signed int overflow UB in 1<<31 constant
Both timestamp-wrap comparisons spelled the "half the u32 range"
sentinel as:

  (unsigned int)(1<<31)

The cast is applied after the shift, so `1<<31` runs in signed int
space and shifts into the sign bit — that's signed integer overflow
and explicit UB under the C/C++ standard. In practice it happens to
produce 0x80000000 on current two's-complement targets, but modern
optimisers are free to assume "no signed overflow" and fold the
comparison the wrong way, which would silently break timestamp
ordering in AmPlayoutBuffer and the wheel timer.

Write the constant as `1U<<31` (unsigned shift) so the value is
well-defined and the cast is no longer needed.

Backport of yeti-switch/sems commit b060cc19 ("fix integerOverflow").
Credit to Igor Ponomarenko / yeti-switch.

+++ core/SampleArray.cc
@@ -37,10 +37,10 @@
 //     return (t1 - t2 > (unsigned int)(1<<31));
 // }
 
-inline bool ts_less::operator()(const unsigned int& l, 
+inline bool ts_less::operator()(const unsigned int& l,
 				const unsigned int& r) const
 {
-  return (l - r > (unsigned int)(1<<31));
+  return (l - r > (1U<<31));
 }
 
 inline bool sys_ts_less::operator()(const unsigned long long& l,
